Extract win caption layout scope to new file
Introduce brave_win_caption_layout.cc containing the thread_local depth and the ScopedWinCaptionLayoutUsesGeometryTabstripOverlap ctor/dtor implementation, and add a static getter to the header. Update BUILD.gn to include the new .cc alongside the header. Remove the duplicate implementation and base/check include from browser_frame_view_win.cc and have it call the new GetCurrentWinCaptionGeometryTabstripOverlapDepth() instead. This centralizes the depth state (thread_local) and avoids multiple-definition/circular-dependency issues while keeping layout selection logic intact.
