+++ browser/ui/BUILD.gn
@@ -760,10 +760,8 @@ source_set("ui") {
       sources += [
         "views/frame/brave_browser_frame_view_win.cc",
         "views/frame/brave_browser_frame_view_win.h",
-
-        # "views/frame/brave_win_caption_layout.h",  # Removed per project policy: no new sources here (see gni_sources.md#circular-dependencies)
-        # Per project circular dependency policy, move this file's source inclusion to the appropriate .gni file as documented:
-        # https://github.com/brave/brave-core/blob/master/docs/gni_sources.md#circular-dependencies
+        "views/frame/brave_win_caption_layout.cc",
+        "views/frame/brave_win_caption_layout.h",
       ]
 
       deps += [ "//chrome/browser:titlebar_config" ]
